Use Environmental Variable to Control Debug Output
If you set the `DEBUG` environmental variable(`export DEBUG=true`), the
build.sh script will now keep the debug settings instead of removing
them.

+++ build/build.sh
@@ -22,8 +22,10 @@ prep() {
     ./replacer.sh ./Outputs/Pencil/content/pencil/common/pencil.js
     ./replacer.sh ./Outputs/Pencil/content/pencil/common/util.js
 
-    rm ./Outputs/Pencil/defaults/preferences/personal.js.xulrunner
-    rm ./Outputs/Pencil/defaults/preferences/debug.js
+    if [ ! $DEBUG ]; then
+        rm ./Outputs/Pencil/defaults/preferences/personal.js.xulrunner
+        rm ./Outputs/Pencil/defaults/preferences/debug.js
+    fi
 }
 
 xpi() {
